+++ src/filters/RevisionFilter.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace craft\webhooks\filters;
+
+use Craft;
+use craft\base\ElementInterface;
+use craft\helpers\ElementHelper;
+
+/**
+ * Filters events based on whether the element is a revision
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 2.1
+ */
+class RevisionFilter extends BaseElementFilter
+{
+    public static function displayName(): string
+    {
+        return Craft::t('webhooks', 'Element is a revision');
+    }
+
+    protected static function checkElement(ElementInterface $element, bool $value): bool
+    {
+        $root = ElementHelper::rootElement($element);
+        return $root->getIsRevision() === $value;
+    }
+}